actions: move build_runtimes logic out of entry
